Продолжаем делить модули
